functions/discord_bot/requirements.txt
discord-interactions
google-cloud-pubsub